+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0](https://hub.docker.com/r/mastermindzh/bw-export/tags)
+
+- Added support for apikey authorization
+  - Skips 2 factor authentication
+
 ## [1.1.1]
 
 Cleaned up the export.sh script from extraneous documentation and a useless empty echo.